@@ -32,7 +32,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -69,7 +69,7 @@ class _MyHomePageState extends State<MyHomePage> {
   double x1 = 0;
   double x2 = 0;
 
-  String qfx;
+  late String qfx;
 
   void selectF(){
     setState(() {
